Fix XmlModel attribute exclusions (#4)
Presently the `XmlModel` method `#to_xml_hash`is not correctly
ignoring attributes defined in `#xml_ignore_attributes`. Whilst this
does not cause problems when making requests to the Agris Web
Services as they seem to be ignored, it does make it more
difficult to compare the results of #to_xml_hash because the results
include class references.

So these changes correct the behavior of `#to_xml_hash` to ignore
the attributes listed in `#xml_ignore_attributes`. 

Includes bonus rubocop and hound configuration.