Option to retrieve changed orders with no details (#12)
Change orders_changed_since call to accept an attribute for
asking Agris for details or not. This supports Otis grabbing all
changed orders without details to speed up calls when a large
number of orders have changed. And to ensure each order is up to
date when processed by grabbing the details then.

needed-by westernmilling/otis#293